Update week1.py
Changed last print statement to white text on black background to not look the same as table. The print statement (and empty input statement) was added to keep result open in console when running compiled file (it would just close the console window when done and the user couldn't see the result).